Improve masterArray
The list of tasks now don't just contain items, they contain objects,
that way each item can have a value of checked that can be either
true or false (false by default). This will be used to implement
the checked functionality. I've also added a category key value pair
in each object in masterArray, which will also be used to add a category
emoji to each title. This makes the array and the code a little bit
more complicated to read, but I really could't find a cleaner way to
add this extra data that I think is necessary in the app.

The code has also been updated to be able to work with the new format
of the array